Automated Research Agent/requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.34
langgraph==0.0.38
duckduckgo-search==5.3.0
python-dotenv==1.0.1
openai>=1.0.0